requirements/base.txt
Django>=1.6,<1.7
South==1.0
argparse
django-autoslug==1.7.2
django-floppyforms==1.1.1
django-gmucas==2.1.1
django-leaflet==0.11.1
django-model-utils==2.0.2
djangorestframework>=2.3
distribute
wsgiref
